used flexbox to make image links behave correctly
When using the column-count property on images that have an anchor
element wrapped around them, the width of the clickable area becomes as
wide as the column instead of the image. I could've made the width of
the image = the width of the column, but it would've stretched the image
to make it lose quality, so i use flexbox.